Proper, tunable via script, heat seaking missles with smoke, fire and explosions. Enemy animations now work. New win32 release. Bug fixes to check if particles were actually allocated.
git-svn-id: http://www.piratestudios.com/pammo/svn@287 fc494e2b-ab4b-0410-a0f4-e0a3d7fb9d21

+++ client/Source/Engine2/enemyLoader.cpp
@@ -249,7 +249,7 @@ bool EnemyLoader::parseWeaponHeatSeaker(char* s)
             column = 8;
             break;
         case 9: // Accuracy
-            data->mAccuracy = atol(s);
+            data->mAccuracy = atof(s);
             break;
         case 10: // Damage
             data->mDamage = atol(s);
